+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- [`useViewportScale`](https://github.com/kripod/standard-hooks/blob/master/src/useViewportScale.ts), [`useViewportScrollCoords`](https://github.com/kripod/standard-hooks/blob/master/src/useViewportScrollCoords.ts) and [`useViewportSize`](https://github.com/kripod/standard-hooks/blob/master/src/useViewportSize.ts) hooks ([#66])
+
 ### Changed
 
 - Improve performance of [`useNetworkAvailability`](https://github.com/kripod/standard-hooks/blob/master/src/useNetworkAvailability.ts) by avoiding array allocation during side effect cleanup
@@ -99,3 +103,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#30]: https://github.com/kripod/standard-hooks/issues/30
 [#31]: https://github.com/kripod/standard-hooks/issues/31
 [#33]: https://github.com/kripod/standard-hooks/issues/33
+[#66]: https://github.com/kripod/standard-hooks/issues/66